refactor: component structure cleanup and renaming
