--- .vite/deps/_metadata.json
@@ -1,8 +1,0 @@
-{
-  "hash": "d2530de4",
-  "configHash": "6ab1bcd7",
-  "lockfileHash": "bf9db59f",
-  "browserHash": "8c7f8873",
-  "optimized": {},
-  "chunks": {}
-}
