+++ src/main.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { createRoot } from 'react-dom/client';
 import '@/styles/global.css';
 import App from './App.tsx';
-import ErrorBoundary from './components/layout/ErrorBoundary.tsx';
-import Fallback from './components/layout/Fallback.tsx';
+import ErrorBoundary from './components/shared/error/ErrorBoundary.tsx';
+import Fallback from './components/shared/error/Fallback.tsx';
 
 const root = document.getElementById('root');
