+++ tests/components/search/Search.test.tsx
@@ -1,9 +1,9 @@
-import Hero from '@/components/hero/Hero';
+import Search from '@/components/search/Search';
 import { render, screen } from '@testing-library/react';
 
 describe('Hero component', () => {
   it('renders initial content', () => {
-    render(<Hero onSearch={jest.fn()} searchQuery=""></Hero>);
+    render(<Search onSearch={jest.fn()} searchQuery="" />);
 
     expect(screen.getByText(/Adventure/i)).toBeInTheDocument();
     expect(screen.getByText(/Track/i)).toBeInTheDocument();
